add unit test for keys
