+++ tests/test_old_compat.py
